requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-environ==0.4.5
django-extensions==3.1.3
idna==3.2
psycopg2==2.9.1
PyJWT==1.7.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==6.63.0
urllib3==1.26.6
